Add FORTRAG90 - a very rudimentary slides compiler
When working on my Git workshops, it is usually a PITA to adapt the
input and output of the command line commands. Having to run them by
hand and copy-pasting the output seriously gets on my nerves (at least
after the 5th time).

This tools allows to put the input in the slides, which is then run as a
subprocess and the output is added to the resulting slides. Avoiding
running the commands by hand.